requirements.txt
Django==1.11.1
django-ckeditor==5.2.2
djangorestframework==3.6.3
olefile==0.44
Pillow==4.1.1
pytz==2017.2
